fix stake chart date range end date
Add time portion to date range end date to include all transactions occurring on that date. Previously no time portion was used in the date creation, so it was defaulting to midnight. This caused the range to not including anything occurring on that end date so charts would never show anything for the last day of month (or last day of year for the yearly chart, but that was not as obvious as the day view).