set pny default stylesheet to dark v2
on very first run of the wallet (when no registry keys exist), the wallet checks which stylesheet to use as the default in case no stylesheet setting has been set.  This setting is the defaulting the default to dark.  (basically you have to set the default in 2 places to chain them correctly on first load...the other commit (4b6cf78) set the default in the other location).