add additional comments regarding code duplication for dumpwallet warning
Added comments to indicate that there is code duplication in this functionality. Existing code was partially duplicating functionality/objects and any attempt to remedy that would have taken more refactoring than appropriate for this small amount of duplication, so the duplication was continued for this enhancement, but this warning should help warn future developers of the duplication.