add retry to IP lookup in masternode install script
Sometimes  icanhazip.com times out and causes coinname.conf file have blanks for IP address. The script ignores this and continues and then the daemon will not start correctly and user has no idea what went wrong.

This commit will add a retry (3 times) to the IP lookup.  If the IP is still unknown, then the script will exit out and inform user to try again in a couple minutes. If it continues they should ask for assistance in discord.

Also removed a coin name branding replacement (coinname-cli) that wasn't needed to be done by developer and changed it to use a script constant instead.